requirements.txt
pandas==0.21.0
progressbar2
requests
bs4
